csv logging for stats

+++ src/utils/game_stats.py
@@ -1,9 +1,12 @@
+import copy
 import json
 import os
 import typing as T
 from eth_typing import Address
 
 from utils import logger
+from utils.price import Prices, wei_to_cra_raw, wei_to_tus_raw
+from crabada.types import IdleGame, Team
 
 
 class GameStats(T.TypedDict):
@@ -21,6 +24,24 @@ class GameStats(T.TypedDict):
     cra_usd: float
     commission_tus: float
     outcome: T.Literal["WIN", "LOSE"]
+
+
+NULL_STATS = GameStats(
+    reinforce1=0.0,
+    reinforce2=0.0,
+    gas_start=0.0,
+    gas_reinforce1=0.0,
+    gas_reinforce2=0.0,
+    gas_close=0.0,
+    game_type="MINE",
+    reward_tus=0.0,
+    reward_cra=0.0,
+    avax_usd=0.0,
+    tus_usd=0.0,
+    cra_usd=0.0,
+    commission_tus=0.0,
+    outcome="WIN",
+)
 
 
 class MineLootGameStats(T.TypedDict):
@@ -67,6 +88,76 @@ NULL_GAME_STATS = LifetimeGameStats(
 )
 
 
+def update_game_stats_after_close(
+    team: Team,
+    mine: IdleGame,
+    lifetime_stats: LifetimeGameStats,
+    game_stats: GameStats,
+    prices: Prices,
+    commission: T.Dict[Address, float],
+) -> None:
+    if mine.get("team_id", "") == team["team_id"]:
+        game_type = "MINE"
+        tus_reward = wei_to_tus_raw(mine.get("miner_tus_reward", 0.0))
+        cra_reward = wei_to_cra_raw(mine.get("miner_cra_reward", 0.0))
+        stats = lifetime_stats[game_type]
+    elif mine.get("attack_team_id", "") == team["team_id"]:
+        game_type = "LOOT"
+        tus_reward = wei_to_tus_raw(mine.get("looter_tus_reward", 0.0))
+        cra_reward = wei_to_cra_raw(mine.get("looter_cra_reward", 0.0))
+        stats = lifetime_stats[game_type]
+    else:
+        logger.print_fail_arrow(f"Failed to detect win/loss for mine {mine['game_id']}")
+        return
+
+    team_id = team["team_id"]
+    if team["team_id"] in game_stats:
+        game_stats[team_id]["reward_tus"] = tus_reward
+        game_stats[team_id]["reward_cra"] = cra_reward
+        game_stats[team_id]["game_type"] = game_type
+
+    if mine.get("winner_team_id", "") == team["team_id"]:
+        stats["game_wins"] += 1
+        if team["team_id"] in game_stats:
+            game_stats[team_id]["outcome"] = "WIN"
+    else:
+        stats["game_losses"] += 1
+        if team["team_id"] in game_stats:
+            game_stats[team_id]["outcome"] = "LOSE"
+
+    stats["game_win_percent"] = (
+        100.0 * float(stats["game_wins"]) / (stats["game_wins"] + stats["game_losses"])
+    )
+
+    stats["tus_gross"] = stats["tus_gross"] + tus_reward
+    stats["cra_gross"] = stats["cra_gross"] + cra_reward
+    stats["tus_net"] = stats["tus_net"] + tus_reward
+    stats["cra_net"] = stats["cra_net"] + cra_reward
+
+    for address, commission in commission.items():
+        commission_tus = tus_reward * (commission / 100.0)
+        commission_cra = cra_reward * (commission / 100.0)
+        # convert cra -> tus and add to tus commission, we dont take direct cra commission
+        commission_tus += prices.cra_to_tus(commission_cra)
+
+        if team["team_id"] in game_stats:
+            game_stats[team_id]["commission_tus"] = commission_tus
+
+        lifetime_stats["commission_tus"][address] = (
+            lifetime_stats["commission_tus"].get(address, 0.0) + commission_tus
+        )
+
+        stats["tus_net"] -= commission_tus
+        stats["cra_net"] -= commission_cra
+
+    if team["team_id"] in game_stats:
+        game_stats[team_id]["avax_usd"] = prices.avax_usd
+        game_stats[team_id]["tus_usd"] = prices.tus_usd
+        game_stats[team_id]["cra_usd"] = prices.cra_usd
+
+    lifetime_stats[game_type] = copy.deepcopy(stats)
+
+
 def update_lifetime_stats_format(game_stats: LifetimeGameStats) -> LifetimeGameStats:
     new_game_stats = LifetimeGameStats()
     new_game_stats["MINE"] = MineLootGameStats()
@@ -102,7 +193,10 @@ def get_game_stats(user: str, log_dir: str) -> T.Dict[str, T.Any]:
         return json.load(infile)
 
 
-def write_game_stats(user: str, log_dir: str, game_stats: LifetimeGameStats) -> None:
+def write_game_stats(user: str, log_dir: str, game_stats: LifetimeGameStats, dry_run=False) -> None:
+    if dry_run:
+        return
+
     game_stats_file = get_lifetime_stats_file(user, log_dir)
     with open(game_stats_file, "w") as outfile:
         json.dump(
